refactor: ErrorResponse에서 custom status code 와 message 반환하도록 수정
