+++ src/main/java/community/whatever/onembackendjava/common/exception/ErrorCode.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 
 public enum ErrorCode {
 
-    NOT_FOUND(HttpStatus.NOT_FOUND, "찾을 수 없음");
+    NOT_FOUND_SHORTEN_URL(HttpStatus.NOT_FOUND, "단축 URL을 찾을 수 없습니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
