+++ src/main/java/community/whatever/onembackendjava/common/exception/ErrorResponse.java
@@ -1,12 +1,12 @@
 package community.whatever.onembackendjava.common.exception;
 
 public record ErrorResponse(
-    int code,
+    String code,
     String message
 ) {
     public static ErrorResponse of(ErrorCode errorCode) {
         return new ErrorResponse(
-            errorCode.getStatus().value(),
+            errorCode.name(),
             errorCode.getMessage()
         );
     }
