+++ src/main/java/community/whatever/onembackendjava/common/exception/GlobalExceptionHandler.java
@@ -1,5 +1,8 @@
 package community.whatever.onembackendjava.common.exception;
 
+import community.whatever.onembackendjava.common.exception.custom.BusinessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
@@ -7,8 +10,10 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(BusinessException.class)
-    protected ErrorResponse handleBusinessException(BusinessException ex) {
-        return ErrorResponse.of(ex.getErrorCode());
+    protected ResponseEntity<ErrorResponse> handleBusinessException(BusinessException e) {
+        ErrorCode errorCode = e.getErrorCode();
+        ErrorResponse errorResponse = ErrorResponse.of(errorCode);
+        return new ResponseEntity<>(errorResponse, HttpStatus.valueOf(errorCode.getStatus().value()));
     }
 
 }
