+++ src/main/java/community/whatever/onembackendjava/service/UrlShortenService.java
@@ -1,7 +1,8 @@
 package community.whatever.onembackendjava.service;
 
-import community.whatever.onembackendjava.common.exception.BusinessException;
+import community.whatever.onembackendjava.common.exception.custom.BusinessException;
 import community.whatever.onembackendjava.common.exception.ErrorCode;
+import community.whatever.onembackendjava.common.exception.custom.NotFoundException;
 import community.whatever.onembackendjava.component.UrlShortener;
 import community.whatever.onembackendjava.repository.UrlShortenRepository;
 import org.springframework.stereotype.Service;
@@ -23,7 +24,7 @@ public class UrlShortenService {
 
     public String getOriginUrlByShortenUrlKey(String shortenUrlKey) {
         return urlShortenRepository.findByShortenUrlKey(shortenUrlKey)
-            .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND));
+            .orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_SHORTEN_URL));
     }
 
 }
